// src/navigation/routes-enum.ts
export enum RoutesEnum {
  WELCOME = 'WELCOME',

  TAB_BAR = 'TAB_BAR',

  CHARACTER = 'Character',
  CHARACTER_FILTER = 'CHARACTER_FILTER',
  CHARACTER_STACK = 'CHARACTER_STACK',
  CHARACTER_DETAIL = 'CHARACTER_DETAIL',

  EPISODE = 'Episode',
  EPISODE_FILTER = 'EPISODE_FILTER',
  EPISODE_STACK = 'EPISODE_STACK',
  EPISODE_DETAIL = 'EPISODE_DETAIL',

  LOCATION = 'Location',
  LOCATION_FILTER = 'LOCATION_FILTER',
  LOCATION_STACK = 'LOCATION_STACK',
  LOCATION_DETAIL = 'LOCATION_DETAIL',
}